+++ src/components/Table.jsx
@@ -1,9 +1,10 @@
-import React from 'react'
+import React, { useEffect } from 'react'
 import { getTableUtils } from '../utils/tableUtils'
 import Root from './Root'
 import { tablePropTypes, reduxEventHandlersPropTypes } from '../types/TableProps'
 import { handlersSymbol } from '../models/Events'
 import defaultTableProps from '../constants/defaultTableProps'
+import * as setUtils from '../utils/setUtils'
 
 /**
  * Table component
@@ -15,13 +16,19 @@ const Table = React.forwardRef((props, ref) => {
   const { namespace, ...rootProps } = props
 
   const utils = getTableUtils(namespace)
+  const { hooks, selectors, events } = utils
 
   // Register redux event handlers
-  const eventHandlers = utils.events[handlersSymbol]
   for (const handlerName in reduxEventHandlersPropTypes) {
-    eventHandlers[handlerName] = rootProps[handlerName]
+    events[handlersSymbol][handlerName] = rootProps[handlerName]
     delete rootProps[handlerName]
   }
+
+  const getState = hooks.useGetState()
+  const selection = hooks.useSelector(selectors.getSelection, setUtils.isEqual)
+  useEffect(() => {
+    events.selectionChange(getState())
+  }, [selection, events, getState])
 
   return <Root
     {...rootProps}